proyecto completo falta ultima revision

+++ app/Http/Controllers/PermissionsController.php
@@ -26,6 +26,9 @@ class PermissionsController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->can('CrearPermiso'))
+            abort(403, 'Permiso Denegado.');
+
         return view('permisos.create');
     }
 
@@ -80,6 +83,9 @@ class PermissionsController extends Controller
      */
     public function edit($id)
     {
+        if (!Auth::user()->can('EditarPermiso'))
+            abort(403, 'Permiso Denegado.');
+
         $permission = Permission::findOrFail($id);
         return view('permisos.edit', ['permission'=>$permission]);
     }
@@ -126,6 +132,9 @@ class PermissionsController extends Controller
      */
     public function destroy($id)
     {
+        if (!Auth::user()->can('EliminarPermiso'))
+            abort(403, 'Permiso Denegado.');
+
         try{
             \DB::beginTransaction();
             Permission::destroy($id);
